refactor(daemon): extract lib.rs unit tests to sibling tests.rs
`crates/uffs-daemon/src/lib.rs` was approaching the 800-LOC soft
limit (729 LOC).  The crate-root `#[cfg(test)] mod tests { ... }`
block (drive_letter_matches contract pinning, ~60 LOC) is moved
into a sibling `src/tests.rs` file, mirroring the pattern already
used by:

  • `crates/uffs-core/src/compact_cache/tests.rs`
  • `crates/uffs-core/src/compact_mmap/tests.rs`
  • `crates/uffs-core/src/compact_storage/tests.rs`
  • `crates/uffs-daemon/src/runtime_orphans/tests.rs`

After extraction `lib.rs` is 670 LOC and `src/tests.rs` is 68 LOC.
Test coverage and module-private visibility (the inline mod was a
plain `mod tests`, the new file resolves to the same submodule
path) are unchanged: `cargo nextest run -p uffs-daemon
drive_letter` reports the same 3/3 pass.

Workspace `cargo clippy --workspace --all-targets --all-features
-- -D warnings` clean.  File-size policy clean.
